The first of probably many commits adding additional SIMD functionality to GLM.
This commit places simdVec4::Data into a nameless union with x, y, z, w components. This will allow .x, etc syntax to be used in the same way as standard vec4/3/2. This requires a warning silencer for MSVC. No preprocessor options controlling this behaviour yet, but will probably come.

Also added is operator[] for simdVec4 which is the same as vec4's.

Finally, the simd_vec4 test program has been updated with basic tests. This will continue to be updated as modifications are made.